+++ app/helpers/database_session.py
@@ -1,4 +1,4 @@
-from app.core.database import SessionLocal
+from app.database import SessionLocal
 from fastapi import Depends
 from sqlalchemy.orm import Session
 from typing import Annotated
